medication-reminder/backend/requirements.txt
Flask
Flask-Cors
pymongo        # if using MongoDB
flask-bcrypt

